Move components to product-editor package and fix UI bug created in checkbox component
