+++ packages/js/product-editor/src/components/checkbox-control/style.scss
