+++ packages/js/product-editor/src/components/index.ts
@@ -39,3 +39,6 @@ export {
 	ManageDownloadLimitsModal as __experimentalManageDownloadLimitsModal,
 	ManageDownloadLimitsModalProps,
 } from './manage-download-limits-modal';
+
+export { Checkbox as __experimentalCheckboxControl } from './checkbox-control';
+export { NumberControl as __experimentalNumberControl } from './number-control';
