+++ packages/js/product-editor/src/hooks/use-number-input-props.ts
